Let Formik manage the count stepper value
N.B. This has the unfortunate side-effect of preventing
react-stepper-primitive from constraining the value to a specified
max. But without this change, refreshing the page will not
reinitialize the initial count value as react-stepper-primitive only
seems to set it the first time, and the value changes after loading
the user's previous value from localStorage in componentDidMount.
Needs further investigation.