refactor(stepper): control own value and remove `focused`
This changes the implementation of the stepper hook to control its own value all the time, rather
than only when unfocused. `useStepper` now ignores the `value` option and `onChange` has been
renamed to `onNewValue` to avoid confusion with the actual `onChange` handler attached to the
input (these are not the same thing and receive different arguments when invoked). This change also
removes the `focused` return value from `useStepper` as there is no need for the hook to track this
information or provide it to the caller.