feat(removed): remove Meetup API key and specific event ID options (#78)
Meetup is removing support for using an API key as of August 15th, 2019. And nobody ever uses the
specific event ID feature anyway. The specific event ID functionality has been removed from the UI
but left in the `draw` Netlify Function in case we want to bring it back. If not, that logic can be
removed in the future as well.

BREAKING CHANGE: Specifying a specific event ID (rather than the soonest upcoming/in-progress event)
is no longer supported. Specifying a Meetup API key (to get full names of winners) is also no longer
supported.

Closes #77.